fix: correct Russian translation for "Servers" and refactor stats display component
