+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "i18n-ally.localesPaths": [
+    "public/locales"
+  ],
+  "i18n-ally.keystyle": "nested"
+}
