+++ public/locales/ru.json
@@ -3,7 +3,7 @@
   "Admins": "Админы",
   "Bans": "Баны",
   "Mutes": "Муты",
-  "Servers": "Serverlar",
+  "Servers": "Серверы",
   "BansAndMute": "Баны и Муты",
   "Skins": "Скины",
   "Rules": "Правила",
